Fixed up a bunch of things all over
Changed session time to 60 minutes etc in case there's bugs with the
carting system (shouldn't be)
added validation to shipping model which is going to share command of
commerce module with cart controller/model. nothing seems particularly
breaky at the moment, this should be ready for launch soon. Need to
create the back end for shipping etc. right now that's not included in
the price calculations.